Updated graph-kubernetes helm chart application dependency to 0.9.2 (updates integration sdk dependencies and updates service account j1 user entity with required "active" field)
Updates the entry docker command to execute (for the cronjob), which has been updated to call the run command (which performs a jupiterone integration collect and sync) to match integration sdk conventions.

Fixes broken release 0.9.1, that required sdk dependencies to first be updated to allow base api url to be customized.

Maintains backwards compatibility